+++ system-parachains/coretime/coretime-kusama/src/tests.rs
@@ -347,3 +347,11 @@ fn dust_accumulates_instead_of_being_burned() {
 			assert_eq!(Balances::total_issuance(), issuance_before);
 		});
 }
+
+#[test]
+fn session_key_deposit_works() {
+	system_parachains_common::test_helpers::session_key_deposit_works::<Runtime>(|owner| {
+		let generated = SessionKeys::generate(&owner.encode(), None);
+		(generated.keys, generated.proof.encode())
+	});
+}